Clean up, refactoring of project structure
