+++ src/card/tests/README.md
@@ -1,0 +1,2 @@
+## Tests
+This folder is intended for tests to be added in the near future.
