+++ src/chart/SingleValueChart.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { ChartProps } from './Chart';
 
-import { renderValueByType } from '../report/RecordProcessing';
+import { renderValueByType } from '../report/ReportRecordProcessing';
 /**
  * Renders Neo4j records as their JSON representation.
  */
