--- src/chart/visualizations/README.md
@@ -1,2 +1,0 @@
-## Neo4j-Labs Charts Visualizations
-The visualizations in this folder were extracted from https://github.com/neo4j-labs/charts.
