+++ src/config/CardConfig.ts
