+++ src/config/DashboardConfig.ts
